Revert "Merge branch 'main' of https://github.com/SimpMahiru/duan1"
This reverts commit f18dc2ac554c743dfe733386136c94ad63c85a7b, reversing
changes made to 924797ce29dfc36533638905c2913fb2fc8d950f.

+++ Quan_Ly_Ban_Ve_May_bay/src/entity/HoaDonCT.java
@@ -13,12 +13,12 @@ public class HoaDonCT {
     private String IDHoaDonCT;
     private String IDVe;
     private String SoLuongVe;
-    private float ThanhTien;
+    private int ThanhTien;
 
     public HoaDonCT() {
     }
 
-    public HoaDonCT(String IDHoaDonCT, String IDVe, String SoLuongVe, float ThanhTien) {
+    public HoaDonCT(String IDHoaDonCT, String IDVe, String SoLuongVe, int ThanhTien) {
         this.IDHoaDonCT = IDHoaDonCT;
         this.IDVe = IDVe;
         this.SoLuongVe = SoLuongVe;
@@ -49,11 +49,11 @@ public class HoaDonCT {
         this.SoLuongVe = SoLuongVe;
     }
 
-    public float getThanhTien() {
+    public int getThanhTien() {
         return ThanhTien;
     }
 
-    public void setThanhTien(float ThanhTien) {
+    public void setThanhTien(int ThanhTien) {
         this.ThanhTien = ThanhTien;
     }
